feat: AuthenticationService에서 access token 얻는 메소드 구현
- header에 Authorization을 가져와 access token을 얻음
