+++ src/main/java/com/sedin/qna/config/WebConfigInterceptors.java
@@ -1,0 +1,21 @@
+package com.sedin.qna.config;
+
+import com.sedin.qna.interceptor.AuthenticationInterceptor;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+public class WebConfigInterceptors implements WebMvcConfigurer {
+
+    private final AuthenticationInterceptor authenticationInterceptor;
+
+    public WebConfigInterceptors(AuthenticationInterceptor authenticationInterceptor) {
+        this.authenticationInterceptor = authenticationInterceptor;
+    }
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(authenticationInterceptor)
+                .excludePathPatterns("/docs")
+                .excludePathPatterns("/api/auth/*");
+    }
+}
