+++ src/main/java/com/sedin/qna/config/WebMvcConfig.java
@@ -1,8 +1,8 @@
 package com.sedin.qna.config;
 
 import com.sedin.qna.interceptor.AuthenticationInterceptor;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
@@ -20,10 +20,8 @@ public class WebMvcConfig implements WebMvcConfigurer {
         registry.addResourceHandler("/api/docs/**").addResourceLocations("classpath:/static/docs/");
     }
 
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(authenticationInterceptor)
-                .excludePathPatterns("/docs")
-                .excludePathPatterns("/api/auth/*");
+    @Bean
+    public WebMvcConfigurer webMvcConfigurerInterceptors() {
+        return new WebConfigInterceptors(authenticationInterceptor);
     }
 }
